+++ src/Menu/AppBar/AppBar.js
@@ -7,7 +7,9 @@ import Toolbar from '@material-ui/core/Toolbar';
 import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
 
-import { mdiChevronRight, mdiMenu } from '@mdi/js';
+import { mdiMenu } from '@mdi/js';
+import NavigateNextIcon from '@material-ui/icons/NavigateNext';
+
 import Icon from '@mdi/react';
 import Link from '@material-ui/core/Link';
 import { DefaultProductTypography } from '../DefaultProductTypography';
@@ -30,6 +32,7 @@ const useStyles = makeStyles(({ palette, spacing }) => ({
     display: 'flex',
     flexGrow: 0,
     position: 'unset',
+    backgroundColor: palette.primary.main,
   },
   breadcrumb: { display: 'flex !important' },
   paperBreadcrumb: {
@@ -43,19 +46,24 @@ const useStyles = makeStyles(({ palette, spacing }) => ({
   link: {
     color: `${palette.common.white} !important`,
   },
+  separator: {
+    color: 'white',
+  },
 }));
 
 export const AppBar = ({
   title,
   subtitle,
-  titleComponent,
+  titleComponent = null,
   menuOnClick,
   leftIcons,
   rightIcons,
+  className,
+  breadcrumbs = [],
 }) => {
   const classes = useStyles();
   return (
-    <div>
+    <div className={className}>
       <MaterialAppBar className={classes.appBar}>
         <Toolbar disableGutters className={classes.toolbar}>
           <IconButton
@@ -68,9 +76,9 @@ export const AppBar = ({
           {leftIcons}
           <div className={classes.grow}>
             {titleComponent ? (
-              <DefaultProductTypography title={title} subtitle={subtitle} />
-            ) : (
               { titleComponent }
+            ) : (
+              <DefaultProductTypography title={title} subtitle={subtitle} />
             )}
           </div>
           {rightIcons}
@@ -78,28 +86,35 @@ export const AppBar = ({
         <div className={classes.paperBreadcrumb}>
           <Breadcrumbs
             classes={{ root: classes.breadcrumb }}
-            separator={<Icon path={mdiChevronRight} color="white" size={1} />}
+            separator={
+              <NavigateNextIcon
+                fontSize="small"
+                className={classes.separator}
+              />
+            }
             arial-label="Breadcrumb"
           >
-            <Link
-              href="/"
-              component="a"
-              variant="subtitle2"
-              classes={{ root: classes.link }}
-            >
-              Portal
-            </Link>
-            <Link
-              href="/"
-              component="a"
-              variant="subtitle2"
-              classes={{ root: classes.link }}
-            >
-              CRM
-            </Link>
-            <Typography color="textPrimary" variant="subtitle2">
-              #1234
-            </Typography>
+            {breadcrumbs.map((current, index, arr) =>
+              index === arr.length - 1 ? (
+                <Typography
+                  key={current.title}
+                  color="textPrimary"
+                  variant="subtitle2"
+                >
+                  {current.title}
+                </Typography>
+              ) : (
+                <Link
+                  key={current.title}
+                  component={current.component}
+                  variant="subtitle2"
+                  classes={{ root: classes.link }}
+                  {...current.componentProps}
+                >
+                  {current.title}
+                </Link>
+              )
+            )}
           </Breadcrumbs>
         </div>
       </MaterialAppBar>
@@ -115,12 +130,19 @@ AppBar.defaultProps = {
   leftIcons: null,
   rightIcons: null,
 };
+const breadcrumb = PropTypes.shape({
+  title: PropTypes.oneOfType([PropTypes.string, PropTypes.func]).isRequired,
+  component: PropTypes.function,
+  componentProps: PropTypes.object,
+});
+
 AppBar.propTypes = {
   title: PropTypes.string,
   subtitle: PropTypes.string,
-  titleComponent: PropTypes.func,
+  titleComponent: PropTypes.object,
   menuOnClick: PropTypes.func,
   leftIcons: PropTypes.object,
   rightIcons: PropTypes.object,
+  breadcrumbs: PropTypes.arrayOf(breadcrumb).isRequired,
 };
 export default AppBar;